Add files via upload
The excel file attached is an example of the data I plan to draw from the original dataset. The original dataset provides dates, opening price, closing price, trade volume, and market cap for each day for various cryptocurrencies. From all of this, I want to narrow down the dataset using R by first looking at all of the information across all dates, find the dates where trade volume reached its top 50%, and then the dates where market cap reached its top 50%, and then find the overlap in these dates. I intend to run this analysis for Bitcoin and Ethereum which are currencies that are more popular and a little more value-based given their utility, and for Dogecoin and Monacoin which are currencies that are considered "meme coins" and have been made popular solely because of social media buzz.